requirements.txt
# Core dependencies
numpy>=1.24.0
pyyaml>=6.0
click>=8.0.0
matplotlib>=3.7.0

# Quantum computing libraries
qiskit>=1.0.0
qiskit-aer>=0.13.0
qiskit-nature>=0.7.0
pennylane>=0.35.0
pennylane-qiskit>=0.35.0

# Classical quantum chemistry (for validation)
pyscf>=2.3.0

# Testing and development
spacy>=3.7.2,<4.0
pytest>=8.0.0
mypy>=1.8.0
flake8>=6.1.0
